# docs/swagger.yaml
definitions:
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
...
  reqmodel.UserListResponse:
    properties:
      user_list:
        items:
          $ref: '#/definitions/sqlc.UserDatum'
        type: array
    type: object
...
  reqmodel.UserUpdateRequest:
    properties:
      login:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  sqlc.UserDatum:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Basic swagger for current api
  termsOfService: http://swagger.io/terms/
  title: movie_backend_go
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Get user list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.UserListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Show user list
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: User creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.UserDatum'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Update user
      tags:
      - user
    post:
      consumes:
      - multipart/form-data
      parameters:
      - in: formData
        name: login
        type: string
      - in: formData
        name: name
        type: string
      - in: formData
        name: password
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sqlc.UserDatum'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Create user
      tags:
      - user
  /user/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete any user as admin
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Delete user
      tags:
      - user
      - admin
    get:
      consumes:
      - application/json
      description: Get user by id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.UserDatum'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Show user
      tags:
      - user
...
  /user/me:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
      - user
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: http://localhost:8080/auth/login
    type: oauth2
swagger: "2.0"